Extract LoadConfig
